+++ guillotina/contrib/cache/serialize.py
@@ -2,15 +2,16 @@ from guillotina.profile import profilable
 
 import asyncpg
 import pickle
+import typing
 
 
 @profilable
-def dumps(value):
+def dumps(value: typing.Any) -> bytes:
     """
-    Serialize the received value using ``json.dumps``.
+    Serialize the received value using ``pickle.dumps``.
 
     :param value: dict
-    :returns: str
+    :returns: bytes
     """
     if isinstance(value, asyncpg.Record):
         value = dict(value)
@@ -18,12 +19,12 @@ def dumps(value):
 
 
 @profilable
-def loads(value):
+def loads(value: bytes) -> typing.Any:
     """
-    Deserialize value using ``ujson.loads``.
+    Deserialize value using ``pickle.loads``.
 
-    :param value: str
-    :returns: output of ``json.loads``.
+    :param value: bytes
+    :returns: output of ``pickle.loads``.
     """
     if value is None:
         return None